+++ bits.h
@@ -24,7 +24,7 @@
 char *strndup(const char *s, size_t size);
 #endif
 
-int wordline(const char *msg, unsigned int x, char **out, int *l, int *i, colour lc); // prepares a string for printing, breaking lines in between words; returns new x
+int wordline(const char *msg, unsigned int x, char **out, size_t *l, size_t *i, colour lc); // prepares a string for printing, breaking lines in between words; returns new x
 void crush(char **buf, unsigned int len);
 void scrush(char **buf, unsigned int len);
 char *mktag(char *fmt, char *from);
